--- src/main/java/com/coloronme/admin/Domain/dto/request/LoginRequestDto.java
@@ -1,19 +1,0 @@
-package com.coloronme.admin.Domain.dto.request;
-
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class LoginRequestDto {
-
-    private String email;
-
-    private String password;
-
-    public LoginRequestDto(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-}
