--- src/main/java/com/coloronme/admin/Domain/dto/response/SignupResponseDto.java
@@ -1,14 +1,0 @@
-package com.coloronme.admin.Domain.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class SignupResponseDto {
-    private String id;
-}
