--- src/main/java/com/coloronme/admin/Domain/entity/Authority.java
@@ -1,5 +1,0 @@
-package com.coloronme.admin.Domain.entity;
-
-public enum Authority {
-    ROLE_USER
-}
