--- src/main/java/com/coloronme/admin/Domain/repository/MemberRepository.java
@@ -1,12 +1,0 @@
-package com.coloronme.admin.Domain.repository;
-
-import com.coloronme.admin.Domain.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByEmail(String email);
-}
