--- src/main/java/com/coloronme/admin/Domain/repository/RefreshTokenRepository.java
@@ -1,13 +1,0 @@
-package com.coloronme.admin.Domain.repository;
-
-import com.coloronme.admin.Domain.entity.RefreshToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-
-    Optional<RefreshToken> findByMemberEmail(String email);
-}
