--- src/main/java/com/coloronme/admin/Global/dto/GlobalResDto.java
@@ -1,16 +1,0 @@
-package com.coloronme.admin.Global.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class GlobalResDto {
-    private String msg;
-    private int statusCode;
-
-    public GlobalResDto(String msg, int statusCode) {
-        this.msg = msg;
-        this.statusCode = statusCode;
-    }
-}
