--- src/main/java/com/coloronme/admin/Global/dto/JwtDto.java
@@ -1,17 +1,0 @@
-package com.coloronme.admin.Global.dto;
-
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class JwtDto {
-    private String accessToken;
-    private String refreshToken;
-
-    public JwtDto(String accessToken, String refreshToken) {
-        this.accessToken = accessToken;
-        this.refreshToken = refreshToken;
-    }
-}
