--- src/main/java/com/coloronme/admin/Global/exception/GlobalExceptionHandler.java
@@ -1,19 +1,0 @@
-package com.coloronme.admin.Global.exception;
-
-
-import com.coloronme.admin.Global.dto.ResponseDto;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(value = {RequestException.class})
-    public ResponseDto<Object> handleApiRequestException(RequestException e) {
-
-        return ResponseDto.fail(
-                e.getHttpStatus(),
-                e.getMessage()
-        );
-    }
-}
