--- src/main/java/com/coloronme/admin/Global/exception/RequestException.java
@@ -1,16 +1,0 @@
-package com.coloronme.admin.Global.exception;
-
-
-import com.coloronme.admin.Global.exception.ErrorCode;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-public class RequestException extends RuntimeException {
-    private final HttpStatus httpStatus;
-
-    public RequestException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.httpStatus = errorCode.getHttpStatus();
-    }
-}
