--- src/main/java/com/coloronme/admin/Global/security/UserDetailsImpl.java
@@ -1,59 +1,0 @@
-package com.coloronme.admin.Global.security;
-
-import com.coloronme.admin.Domain.entity.Member;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-
-public class UserDetailsImpl implements UserDetails {
-
-    public Member member;
-
-    public UserDetailsImpl(Member member) {
-        this.member = member;
-    }
-
-    public Member getMember() {
-        return this.getMember();
-    }
-
-    public void setMember(Member member) {
-        this.member = member;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return null;
-    }
-
-    @Override
-    public String getPassword() {
-        return null;
-    }
-    @Override
-    public String getUsername() {
-        return null;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return false;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return false;
-    }
-}
-
