--- src/main/java/com/coloronme/admin/Global/security/UserDetailsServiceImpl.java
@@ -1,40 +1,0 @@
-package com.coloronme.admin.Global.security;
-
-import com.coloronme.admin.Domain.repository.MemberRepository;
-import com.coloronme.admin.Domain.entity.Member;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.Collections;
-
-@Service
-@RequiredArgsConstructor
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-    private final MemberRepository memberRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-
-        Member member = memberRepository.findByEmail(email).
-                orElseThrow(() -> new RuntimeException("Not Found Account")
-                );
-        return createUserDetails(member);
-    }
-
-    private UserDetails createUserDetails(Member member) {
-        GrantedAuthority grantedAuthority = new SimpleGrantedAuthority("member");
-
-        return new User(
-                member.getEmail(),
-                member.getPassword(),
-                Collections.singleton(grantedAuthority)
-        );
-    }
-}
