+++ src/main/java/com/coloronme/admin/domain/consultant/dto/response/LoginResponseDto.java
@@ -1,0 +1,19 @@
+package com.coloronme.admin.domain.consultant.dto.response;
+
+
+import com.coloronme.admin.domain.consultant.entity.Authority;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class LoginResponseDto {
+
+    private String email;
+    private Authority authority;
+
+}
