+++ src/main/java/com/coloronme/admin/domain/consultant/dto/response/SignupResponseDto.java
@@ -1,0 +1,14 @@
+package com.coloronme.admin.domain.consultant.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class SignupResponseDto {
+    private String id;
+}
