+++ src/main/java/com/coloronme/admin/domain/consultant/entity/Authority.java
@@ -1,0 +1,5 @@
+package com.coloronme.admin.domain.consultant.entity;
+
+public enum Authority {
+    ROLE_USER
+}
