+++ src/main/java/com/coloronme/admin/global/HomeController.java
@@ -1,0 +1,12 @@
+package com.coloronme.admin.global;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HomeController {
+    @GetMapping("/health")
+    public String checkHealth(){
+        return "Connection Status is Ok.";
+    }
+}
