+++ src/main/java/com/coloronme/admin/global/dto/Error.java
@@ -1,0 +1,18 @@
+package com.coloronme.admin.global.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class Error {
+
+    private HttpStatus httpStatus;
+
+    private String message;
+}
