+++ src/main/java/com/coloronme/admin/global/dto/GlobalResDto.java
@@ -1,0 +1,16 @@
+package com.coloronme.admin.global.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class GlobalResDto {
+    private String msg;
+    private int statusCode;
+
+    public GlobalResDto(String msg, int statusCode) {
+        this.msg = msg;
+        this.statusCode = statusCode;
+    }
+}
