+++ src/main/java/com/coloronme/admin/global/dto/JwtDto.java
@@ -1,0 +1,17 @@
+package com.coloronme.admin.global.dto;
+
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class JwtDto {
+    private String accessToken;
+    private String refreshToken;
+
+    public JwtDto(String accessToken, String refreshToken) {
+        this.accessToken = accessToken;
+        this.refreshToken = refreshToken;
+    }
+}
