+++ src/main/java/com/coloronme/admin/global/exception/GlobalExceptionHandler.java
@@ -1,0 +1,19 @@
+package com.coloronme.admin.global.exception;
+
+
+import com.coloronme.admin.global.dto.ResponseDto;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(value = {RequestException.class})
+    public ResponseDto<Object> handleApiRequestException(RequestException e) {
+
+        return ResponseDto.fail(
+                e.getHttpStatus(),
+                e.getMessage()
+        );
+    }
+}
