+++ src/main/java/com/coloronme/admin/global/exception/RequestException.java
@@ -1,0 +1,15 @@
+package com.coloronme.admin.global.exception;
+
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+public class RequestException extends RuntimeException {
+    private final HttpStatus httpStatus;
+
+    public RequestException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.httpStatus = errorCode.getHttpStatus();
+    }
+}
