+++ src/main/java/com/coloronme/admin/global/security/UserDetailsImpl.java
@@ -1,0 +1,59 @@
+package com.coloronme.admin.global.security;
+
+import com.coloronme.admin.domain.consultant.entity.Consultant;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+
+public class UserDetailsImpl implements UserDetails {
+
+    public Consultant member;
+
+    public UserDetailsImpl(Consultant member) {
+        this.member = member;
+    }
+
+    public Consultant getMember() {
+        return this.getMember();
+    }
+
+    public void setMember(Consultant member) {
+        this.member = member;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return null;
+    }
+
+    @Override
+    public String getPassword() {
+        return null;
+    }
+    @Override
+    public String getUsername() {
+        return null;
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return false;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return false;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return false;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return false;
+    }
+}
+
