+++ src/main/java/com/coloronme/admin/global/security/UserDetailsServiceImpl.java
@@ -1,0 +1,40 @@
+package com.coloronme.admin.global.security;
+
+import com.coloronme.admin.domain.consultant.repository.ConsultantRepository;
+import com.coloronme.admin.domain.consultant.entity.Consultant;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import java.util.Collections;
+
+@Service
+@RequiredArgsConstructor
+public class UserDetailsServiceImpl implements UserDetailsService {
+
+    private final ConsultantRepository consultantRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+
+        Consultant member = consultantRepository.findByEmail(email).
+                orElseThrow(() -> new RuntimeException("Not Found Account")
+                );
+        return createUserDetails(member);
+    }
+
+    private UserDetails createUserDetails(Consultant member) {
+        GrantedAuthority grantedAuthority = new SimpleGrantedAuthority("member");
+
+        return new User(
+                member.getEmail(),
+                member.getPassword(),
+                Collections.singleton(grantedAuthority)
+        );
+    }
+}
